CVS_CO_DIR may be empty now.

+++ builder_scripts/builder_common.sh
@@ -1642,10 +1642,12 @@ pfSense_clean_obj_dir() {
 	freesbie_make cleandir
 	echo -n "Cleaning up previous build environment...Please wait..."
 	# Allow old CVS_CO_DIR to be deleted later
-	if [ -d $CVS_CO_DIR ]; then 
-		echo -n "."
-		chflags -R noschg $CVS_CO_DIR
-		rm -rf $CVS_CO_DIR
+	if "$CVS_CO_DIR" != "" ]; then
+		if [ -d $CVS_CO_DIR ]; then 
+			echo -n "."
+			chflags -R noschg $CVS_CO_DIR
+			rm -rf $CVS_CO_DIR
+		fi
 	fi
 	echo -n "."
 	if [ -d "${PFSENSEBASEDIR}/dev" ]; then
